requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
python-dotenv==1.0.0
crewai==0.152.0
gunicorn==21.2.0
requests==2.31.0
